feat(store): manage default attr set, group, code creation

+++ apps/server/src/attribute/attribute.service.ts
@@ -82,17 +82,19 @@ export class AttributeService {
       ...rest
     } = updateAttributeDto;
 
-    const productTypeCreate: Prisma.product_type_attributeUncheckedCreateWithoutAttributeInput[] = productTypes.map((productTypeId) => ({ product_type_id: productTypeId }));
-
-    await this.repository.deleteAllProductTypes(id);
-    await this.repository.deleteAllOptions(id);
+    if (productTypes) await this.repository.deleteAllProductTypes(id);
+    if (options) await this.repository.deleteAllOptions(id);
 
     return this.repository.update({
       where: { id },
       data: {
         ...rest,
-        product_type_attribute: { create: productTypeCreate },
-        attribute_option: { create: options },
+        ...(productTypes && {
+          product_type_attribute: {
+            create: productTypes.map((productTypeId) => ({ product_type_id: productTypeId })),
+          },
+        }),
+        ...(options && { attribute_option: { create: options } }),
       },
     });
   }
